+++ boot.php
@@ -9,7 +9,7 @@ if (rex_addon::exists('yform') &&
     rex_addon::get('yform')->isAvailable() &&
     rex_plugin::get('yform', 'manager')->isAvailable()) {
     rex_yform::addTemplatePath(rex_path::addon('mform', 'ytemplates'));
-    rex_extension::register('MEDIA_IS_IN_USE', 'MformYformHelper::isMediaInUse');
+    rex_extension::register('MEDIA_IS_IN_USE', "FriendsOfRedaxo\\MformYformHelper::isMediaInUse");
 }
 
 if (rex::isBackend()) {