+++ lib/MForm/Handler/MFormValueHandler.php
@@ -5,10 +5,10 @@
  * @license MIT
  */
 
-namespace MForm\Handler;
+namespace FriendsOfRedaxo\MForm\Handler;
 
-use MForm\DTO\MFormItem;
-use MForm\Utils\MFormClang;
+use FriendsOfRedaxo\MForm\DTO\MFormItem;
+use FriendsOfRedaxo\MForm\Utils\MFormClang;
 use rex;
 use rex_logger;
 use rex_sql;