+++ lib/MForm/Inputs/MFormInputsInterface.php
@@ -5,10 +5,10 @@
  * @license MIT
  */
 
-namespace MForm\Inputs;
+namespace FriendsOfRedaxo\MForm\Inputs;
 
-use MForm;
-use MForm\DTO\MFormInputsConfig;
+use FriendsOfRedaxo\MForm;
+use FriendsOfRedaxo\MForm\DTO\MFormInputsConfig;
 
 interface MFormInputsInterface
 {