+++ pages/demo.demo_community.php
@@ -5,7 +5,7 @@
  * @license MIT
  */
 
-use MForm\Utils\MFormPageHelper;
+use FriendsOfRedaxo\MForm\Utils\MFormPageHelper;
 
 echo rex_view::title(rex_i18n::msg('mform_title') . ' ' . rex_i18n::msg('mform_' . rex_be_controller::getCurrentPagePart(2)));